fix: 챌린지 관리자 삭제 경로 수정
